requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.6.5
amqp==5.1.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
blinker==1.5
celery==5.1.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.1.3
Faker==8.12.0
fastapi==0.79.0
fastapi-mail==1.1.4
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httptools==0.4.0
httpx==0.19.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==0.13.1
prompt-toolkit==3.0.30
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pytest==6.2.4
pytest-asyncio==0.15.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==3.5.3
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
vine==5.0.0
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3
